backend/orchestrator/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.25.2
paho-mqtt==1.6.1
pytest==7.4.4
pytest-asyncio==0.23.2
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx[http2]==0.25.2
